Шаблонизируй то!
Часть 1:

- Переписал модули на функцию
- Пердаю новым модулям статические данные для рендеринга
